+++ CHANGELOG.md
@@ -12,7 +12,7 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 
 ### Changed
 
-- Added new param `skipLagCompensation` to `SWEP:SetNextPrimaryFire(nextTime, skipLagCompensation)` to skip our inbuilt lag compensation
+- Added new param `skipTickrateFix` to `SWEP:SetNextPrimaryFire(nextTime, skipTickrateFix)` to skip our inbuilt tickrate fix
 
 ## [v0.10.2b](https://github.com/TTT-2/TTT2/tree/v0.10.2b) (2021-10-21)